Remove Path trait
